+++ pythonmysequel/Row.py
@@ -10,7 +10,7 @@ class Row:
         self.values = values
 
         self._check_value_type()
-    
+
     def _check_value_type(self):
         for value_type, value, column_name in zip(self.table.values.values(), self.values.values(), self.values):
             inputted_value_type = type(value)
@@ -18,3 +18,7 @@ class Row:
 
             if inputted_value_type !=  python_value_type:
                 raise ValueError(f'Incorrect value type {inputted_value_type} for column "{column_name}" {python_value_type}')
+
+    def _add_value(self, value:dict):
+        for k, v in value.items():
+            self.values[k] = v