+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes will be documented in this file.
 - Ability to create, and use databases
 - Ability to create and drop tables
 - Ability to insert into table
+- Ability to query table
 ### Todo:
 - [ ] Add ability to Row class:
-  - [ ] Interacting with rows (update, delete, query)
+  - [ ] Interacting with rows (update, delete)